Fix NRPT unqualified name
The OpenVPN 3 NRPT implementations make use of
openvpn::Win::check_reg_error() but do not include the namespace
qualification.  Obviously wherever this file is being used includes a
`using` statement, or is within openvpn::Win, and this error was not
detected by OpenVPN.

UH VPN makes use of this file from a different namespace without any
`using` statements and thus require this call to be fully qualified.