Add TUN_PREFIX_WRITE_ONLY compile option
When writing from OpenVPN -> TUN abstraction, NEPacketTunnelFlow on
Apple platforms the protocol is needed, so this is added as a prefix.
For consistency this is also expected when packets are read from the TUN
interface, however for some implementations it can be beneficial to
avoid this for performance gains.

As the prefix is just skipped when reading from TUN -> OpenVPN, this
patch adds an option that makes the tun_prefix option work in an
asymmetric manner, only affecting packets that are written to the TUN
interface.