options: Fix incorrect argument count in min_args()
The data vector contains the option name on the first element (0)
and the arguments to the option in the following arguments.  And
when calling min_args() where the option is lacking arguments,
there is an off-by-one error in the error message.

There are not many callers of min_args(), and those places it was
used this seems to correct a similar issue everywhere.

This was first detected by having '--keepalive 30' in a
configuration profile, where the exception caught produced this
error:

    option_error: option 'keepalive' must have at least 3 arguments

This is wrong, as --keepalive needs at least 2 arguments.

Signed-off-by: David Sommerseth <davids@openvpn.net>